apps/analytics/requirements.txt
# Analytics and Machine Learning Dependencies
# Core ML libraries
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=1.0.0

# Deep Learning (optional)
tensorflow>=2.8.0
torch>=1.11.0

# Data Visualization
matplotlib>=3.5.0
seaborn>=0.11.0
plotly>=5.0.0

# Time Series Analysis
statsmodels>=0.13.0
prophet>=1.0.0

# Clustering and Dimensionality Reduction
umap-learn>=0.5.0
hdbscan>=0.8.0

# Recommendation Systems
surprise>=1.1.0
implicit>=0.6.0

# Feature Engineering
feature-engine>=1.2.0
category-encoders>=2.3.0

# Model Evaluation
mlxtend>=0.19.0
yellowbrick>=1.4.0

# Data Processing
dask>=2022.1.0
joblib>=1.1.0

# Caching and Performance
redis>=4.0.0
django-redis>=5.0.0

# Monitoring and Logging
prometheus-client>=0.12.0
grafana-api>=1.0.0

# Testing
pytest>=6.0.0
pytest-django>=4.4.0
pytest-cov>=3.0.0

# Development
jupyter>=1.0.0
ipython>=7.0.0

